Update brick device details as part of sync.
This patch gathers the gluster brick device details and
updates it during the sds sync. It uses blivet to gather
the node's storage device details.

Using the brick path the brick mount-point and filesystem
is gathered first. using these details we search for the
brick device in blivet's devicetree. Blivet's device tree
will have the details about the underlying device.

tendrl-bug-id: Tendrl/gluster-integration#333
tendrl-bug-id: Tendrl/specifications#173
Signed-off-by: nnDarshan <darshan.n.2024@gmail.com>