Updated to mark the volume state as down if stopped
If the volume status is stopped we mark the state as down rightaway.
Also not all the bricks are reported yet in central store of tendrl
from all the participating nodes of the cluster (i.e. the total no of
bricks reported till now is less than the brick_count field of volume),
the volume state remains as `unknown`.

So now the calculation logic of volume state triggers only if volume
is started and if not all the bricks updated yet in central store, the
volume state is marked as `unknown`.

tendrl-bug-id: Tendrl/monitoring-integration#23
Signed-off-by: Shubhendu <shtripat@redhat.com>